+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2025-04-14
+
+### Added
+
+- Adds support for the `<pre>` element using the `Goatee.Preformatted()` class.
+- Adds support for the `<nav>` element using the `Goatee.Nav()` class.
+
 ## [0.4.1] - 2025-04-14
 
 ### Changes